requirements.txt
# requirements.txt
# ----------------
# Runtime dependencies for AI Project-Analyser

beeai>=0.2.0
rich>=12.0
pyyaml>=5.4
fastapi>=0.78
uvicorn[standard]>=0.17
python-multipart>=0.0.5
jinja2>=3.0
sse-starlette>=1.0
